+++ react/components/RenderProvider.tsx
@@ -243,6 +243,7 @@ export class RenderProvider extends Component<
       components,
       exposeBindingAddress,
       extensions,
+      isJanusProxied,
       pages,
       page,
       query,
@@ -253,7 +254,7 @@ export class RenderProvider extends Component<
       loadedDevices,
     } = props.runtime
     const { apollo, history, deviceInfo, sessionPromise } = props
-    const ignoreCanonicalReplacement = query && query.map
+    const ignoreCanonicalReplacement = isJanusProxied || (query && query.map)
     this.fetcher = fetch
 
     if (binding && canUseDOM) {
@@ -581,13 +582,21 @@ export class RenderProvider extends Component<
   }
 
   private updateDeviceBlocks = async (deviceInfo: DeviceInfo) => {
-    const query = queryStringToMap(location.search) as RenderRuntime['query']
+    const {
+      runtime: { isJanusProxied },
+    } = this.props
 
+    const {
+      route: { path },
+    } = this.state
+
+    const query = queryStringToMap(location.search) as RenderRuntime['query']
     const { components, extensions, messages } = await fetchServerPage({
       fetcher: this.fetcher,
-      path: this.state.route.path,
+      path,
       query,
       deviceInfo,
+      isJanusProxied,
     })
 
     await this.fetchComponents(components, extensions)
@@ -647,7 +656,7 @@ export class RenderProvider extends Component<
 
   public onPageChanged = (location: RenderHistoryLocation) => {
     const {
-      runtime: { renderMajor, query: queryFromRuntime },
+      runtime: { renderMajor, query: queryFromRuntime, isJanusProxied },
     } = this.props
 
     const {
@@ -765,6 +774,7 @@ export class RenderProvider extends Component<
           query,
           workspace: workspaceFromQuery,
           deviceInfo,
+          isJanusProxied,
         }).then(
           async ({
             appsEtag,
@@ -971,7 +981,7 @@ export class RenderProvider extends Component<
 
   public updateRuntime = async (options?: PageContextOptions) => {
     const {
-      runtime: { renderMajor, query: queryFromRuntime },
+      runtime: { renderMajor, query: queryFromRuntime, isJanusProxied },
     } = this.props
     const {
       page,
@@ -1004,6 +1014,7 @@ export class RenderProvider extends Component<
           fetcher: this.fetcher,
           workspace: workspaceFromQuery,
           deviceInfo,
+          isJanusProxied,
         })
       : await fetchNavigationPage({
           apolloClient: this.apolloClient,